requirements.txt
dvc
matplotlib
numpy==1.19.2
pandas
pillow
requests
scikit-image
scikit-learn
tensorflow==2.4.0
tensorflow-addons
torch
tqdm